moved containers to lib folder to allow exporting of components, App.js test fails
