+++ src/lib/containers/UserFavorites.js
