+++ src/lib/index.js
@@ -1,4 +1,9 @@
 import * as SynapseClient from './utils/SynapseClient';
 import * as SynapseConstants from './utils/SynapseConstants';
+import Markdown from './containers/Markdown'
 
-export { SynapseClient, SynapseConstants };
+let SynapseComponents = {
+    Markdown
+}
+
+export { SynapseClient, SynapseConstants, SynapseComponents };
