+++ src/test/lib/utils/Login.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Login from 'demo/containers/Login';
+import Login from 'lib/containers/Login';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
